Refactor spin.ts
