+++ packages/cli-kit/package.json
@@ -70,7 +70,7 @@
                 "./error-handler.js",
                 "../../environment/local.js",
                 "@oclif/core",
-                "./private/node/environment/utilities.js",
+                "../../private/node/environment/utilities.js",
                 "../../constants.js",
                 "../../path.js",
                 "./system.js"
