+++ packages/cli-kit/src/constants.ts
@@ -21,10 +21,7 @@ const constants = {
     runAsUser: 'SHOPIFY_RUN_AS_USER',
     serviceEnv: 'SHOPIFY_SERVICE_ENV',
     skipCliRedirect: 'SHOPIFY_CLI_SKIP_CLI_REDIRECT',
-    spinHost: 'SPIN_HOST',
     spinInstance: 'SPIN_INSTANCE',
-    spinNamespace: 'SPIN_NAMESPACE',
-    spinWorkspace: 'SPIN_WORKSPACE',
     themeToken: 'SHOPIFY_CLI_THEME_TOKEN',
     unitTest: 'SHOPIFY_UNIT_TEST',
     verbose: 'SHOPIFY_FLAG_VERBOSE',
