+++ packages/cli-kit/src/environment.ts
@@ -1,7 +1,5 @@
 import * as local from './environment/local.js'
 import * as service from './environment/service.js'
 import * as fqdn from './environment/fqdn.js'
-import * as spin from './environment/spin.js'
-import * as network from './network/service.js'
 
-export {local, service, fqdn, spin, network}
+export {local, service, fqdn}
