+++ packages/cli-kit/src/environment/local.ts
@@ -1,8 +1,8 @@
-import {isSpin} from './spin.js'
 import {isTruthy, isSet} from '../private/node/environment/utilities.js'
 import constants from '../constants.js'
 import {exists as fileExists} from '../file.js'
 import {exec} from '../public/node/system.js'
+import {isSpin} from '../public/node/environment/spin.js'
 import isInteractive from 'is-interactive'
 import macaddress from 'macaddress'
 import {homedir} from 'os'
