+++ packages/cli-kit/src/private/node/session/identity.ts
@@ -1,7 +1,6 @@
 import {Bug} from '../../../error.js'
-import {Environment} from '../../../network/service.js'
 import {API} from '../../../network/api.js'
-import {serviceEnvironment} from '../../../environment/service.js'
+import {Environment, serviceEnvironment} from '../../../environment/service.js'
 
 export function clientId(): string {
   const environment = serviceEnvironment()
