+++ packages/cli-kit/src/public/node/environment/spin.test.ts
@@ -1,10 +1,10 @@
-import {show, fqdn, isSpin, instance, workspace, namespace, host} from './spin.js'
-import {getCachedSpinFqdn, setCachedSpinFqdn} from './spin-cache.js'
-import {captureOutput} from '../public/node/system.js'
+import {show, isSpin, spinFqdn, instance} from './spin.js'
+import {getCachedSpinFqdn, setCachedSpinFqdn} from '../../../private/node/environment/spin-cache.js'
+import {captureOutput} from '../system.js'
 import {describe, test, expect, vi, it} from 'vitest'
 
-vi.mock('../public/node/system.js')
-vi.mock('./spin-cache')
+vi.mock('../system.js')
+vi.mock('../../../private/node/environment/spin-cache.js')
 
 const mockedCaptureOutput = vi.mocked(captureOutput)
 
@@ -17,7 +17,7 @@ describe('fqdn', () => {
     mockedCaptureOutput.mockResolvedValue(JSON.stringify(showResponse))
 
     // When
-    const got = await fqdn(env)
+    const got = await spinFqdn(env)
 
     // Then
     expect(got).toEqual('fqdn')
@@ -33,7 +33,7 @@ describe('fqdn', () => {
     mockedCaptureOutput.mockResolvedValue(JSON.stringify(showResponse))
 
     // When
-    const got = await fqdn(env)
+    const got = await spinFqdn(env)
 
     // Then
     expect(got).toEqual('fqdn')
@@ -47,7 +47,7 @@ describe('fqdn', () => {
     vi.mocked(getCachedSpinFqdn).mockReturnValue('cachedFQDN')
 
     // When
-    const got = await fqdn(env)
+    const got = await spinFqdn(env)
 
     // Then
     expect(got).toEqual('cachedFQDN')
@@ -129,81 +129,6 @@ describe('instance', () => {
 
     // When
     const got = instance(env)
-
-    // Then
-    expect(got).toBeUndefined()
-  })
-})
-
-describe('workspace', () => {
-  test('returns the value of SPIN_WORKSPACE', () => {
-    // Given
-    const workspaceName = 'workspace'
-    const env = {SPIN_WORKSPACE: workspaceName}
-
-    // When
-    const got = workspace(env)
-
-    // Then
-    expect(got).toBe(workspaceName)
-  })
-
-  test('returns undefined value when SPIN_WORKSPACE is not defined', () => {
-    // Given
-    const env = {}
-
-    // When
-    const got = workspace(env)
-
-    // Then
-    expect(got).toBeUndefined()
-  })
-})
-
-describe('namespace', () => {
-  test('returns the value of SPIN_NAMESPACE', () => {
-    // Given
-    const namespaceName = 'namespace'
-    const env = {SPIN_NAMESPACE: namespaceName}
-
-    // When
-    const got = namespace(env)
-
-    // Then
-    expect(got).toBe(namespaceName)
-  })
-
-  test('returns undefined value when SPIN_NAMESPACE is not defined', () => {
-    // Given
-    const env = {}
-
-    // When
-    const got = namespace(env)
-
-    // Then
-    expect(got).toBeUndefined()
-  })
-})
-
-describe('host', () => {
-  test('returns the value of SPIN_HOST', () => {
-    // Given
-    const hostName = 'host'
-    const env = {SPIN_HOST: hostName}
-
-    // When
-    const got = host(env)
-
-    // Then
-    expect(got).toBe(hostName)
-  })
-
-  test('returns undefined value when SPIN_HOST is not defined', () => {
-    // Given
-    const env = {}
-
-    // When
-    const got = host(env)
 
     // Then
     expect(got).toBeUndefined()
