+++ packages/cli-kit/src/public/node/environment/spin.ts
@@ -1,12 +1,12 @@
-import {getCachedSpinFqdn, setCachedSpinFqdn} from './spin-cache.js'
-import {isTruthy} from '../private/node/environment/utilities.js'
-import constants from '../constants.js'
-import {captureOutput} from '../public/node/system.js'
-import {content, token} from '../output.js'
-import {exists, readSync} from '../file.js'
-import {Abort} from '../error.js'
+import {isTruthy} from '../../../private/node/environment/utilities.js'
+import constants from '../../../constants.js'
+import {captureOutput} from '../system.js'
+import {content, token} from '../../../output.js'
+import {exists, readSync} from '../../../file.js'
+import {getCachedSpinFqdn, setCachedSpinFqdn} from '../../../private/node/environment/spin-cache.js'
+import {Abort} from '../../../error.js'
 
-export const SpinInstanceNotFoundMessages = (spinInstance: string | undefined, error: string) => {
+const SpinInstanceNotFoundMessages = (spinInstance: string | undefined, error: string) => {
   const errorMessage = content`${token.genericShellCommand(
     `spin`,
   )} yielded the following error trying to obtain the fully qualified domain name of the Spin instance:
@@ -25,10 +25,12 @@ const spinFqdnFilePath = '/etc/spin/machine/fqdn'
  * When ran in a Spin environment, it returns the fqdn of the instance.
  *
  * Will cache the value of the Spin FQDN during the execution of the CLI.
- * To avoid multiple calls to `readSync` or `show`
- * @returns fqdn of the Spin environment.
+ * To avoid multiple calls to `readSync` or `show`.
+ *
+ * @param env - Environment variables.
+ * @returns Fqdn of the Spin environment.
  */
-export async function fqdn(env = process.env): Promise<string> {
+export async function spinFqdn(env = process.env): Promise<string> {
   let spinFqdn = getCachedSpinFqdn()
   if (spinFqdn) return spinFqdn
 
@@ -46,7 +48,9 @@ export async function fqdn(env = process.env): Promise<string> {
 
 /**
  * Runs "spin show" and returns the JSON-parsed output.
+ *
  * @param spinInstance - When it's undefined, we'll fetch the latest one.
+ * @param env - Environment variables.
  * @returns The JSON-parsed output of the Spin CLI.
  * @throws Any error raised from the underlying Spin CLI.
  */
@@ -65,7 +69,8 @@ export async function show(spinInstance: string | undefined, env = process.env):
 
 /**
  * Returns true if the CLI is running in a Spin environment.
- * @param env - Environment variables
+ *
+ * @param env - Environment variables.
  * @returns True if the CLI is running in a Spin environment.
  */
 export function isSpin(env = process.env): boolean {
@@ -74,36 +79,10 @@ export function isSpin(env = process.env): boolean {
 
 /**
  * Returns the value of the SPIN_INSTANCE environment variable.
- * @param env - Environment variables
+ *
+ * @param env - Environment variables.
  * @returns The value of the SPIN_INSTANCE environment variable.
  */
 export function instance(env = process.env): string | undefined {
   return env[constants.environmentVariables.spinInstance]
-}
-
-/**
- * Returns the value of the SPIN_WORKSPACE environment variable.
- * @param env - Environment variables
- * @returns The value of the SPIN_WORKSPACE environment variable.
- */
-export function workspace(env = process.env): string | undefined {
-  return env[constants.environmentVariables.spinWorkspace]
-}
-
-/**
- * Returns the value of the SPIN_NAMESPACE environment variable.
- * @param env - Environment variables
- * @returns The value of the SPIN_NAMESPACE environment variable.
- */
-export function namespace(env = process.env): string | undefined {
-  return env[constants.environmentVariables.spinNamespace]
-}
-
-/**
- * Returns the value of the SPIN_HOST environment variable.
- * @param env - Environment variables
- * @returns The value of the SPIN_HOST environment variable.
- */
-export function host(env = process.env): string | undefined {
-  return env[constants.environmentVariables.spinHost]
 }
